Fixed and completed continued text
The idea to save the reference of the message box object doesn't
work in c++: Reference object members must be set in the
constructor and cannot be changed. They are unlike object
references in other languages in this way.

The solution was to fall back to using pointers.

Determining the continuation text is performed by using the last
line printed to the screen as a needle in a string search, then
using the found position to calculate where to continue with the
remaining substring.

The message is replaced, so we can continue long texts as much as
it takes to get to the end.